requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
bleach==6.0.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.31.0
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.3.0
django-summernote==0.8.20.0
django-timezone-field==5.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
twilio==7.16.2
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.3.0
Pillow==8.4.0